restructered code
