+++ cmd/armv/types/errorresponse.go
@@ -21,14 +21,16 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package app
+package types
 
-const (
-	//API return codes
-	API_SUCCESS            int = 202
-	API_RESOURCE_MOVE_OK   int = 204
-	API_RESOURCE_MOVE_FAIL int = 409
-
-	//Progress bar Max
-	PROGRESS_BAR_MAX int = 100
-)
+type ErrorResponse struct {
+	Error struct {
+		Code    string `json:"code"`
+		Message string `json:"message"`
+		Details []struct {
+			Code    string `json:"code"`
+			Target  string `json:"target"`
+			Message string `json:"message"`
+		} `json:"details"`
+	} `json:"error"`
+}
